Removing broken duplicate take_n_random_slices
The function def take_n_random_slices does not take random slices because the orientations (random or not) need to be provided.
This function is broken (length instead of len, quat2rot3d instead of quaternion2rot3d).
This function is an almost direct duplicate of take_n_slice.

I tested them to be equivalent if (corrected for the two broken elements) and called with the following:
    ten_slices = ps.geometry.take_n_slice(
        pattern_shape=det.pedestal.shape,
        pixel_momentum=det.pixel_position_reciprocal,
        volume=volume,
        voxel_length=voxel_length,
        orientations=orientation)
    ten_slices_ = ps.geometry.take_n_random_slices(
        detector=det,
        volume=volume,
        voxel_length=voxel_length,
        orientations=orientation)